backend/API_projects/requirements.txt
fastapi==0.105.0
uvicorn==0.24.0
python-dotenv==1.0.0
supabase==2.0.3
openai==1.3.5
mistralai==0.0.9
pydantic==2.5.2
python-multipart==0.0.6 